[+] Added ability to request for auto release of some of a memory pool's memory if the different in used size and capacity is too big [U] Turned off auto-resize for GameObject pool [U] Memory pools moved to Common/Memory

+++ Assets/Engine/Scripts/Common/Memory/ObjectPool.cs
@@ -6,22 +6,25 @@ namespace Assets.Engine.Scripts.Common.Collections
 {
     public sealed class ObjectPool<T> where T: class
     {
-        // Delegate handling allocation of memory
+        //! Delegate handling allocation of memory
         private readonly ObjectPoolAllocator<T> m_objectAllocator;
-        // Delegate handling releasing of memory
+        //! Delegate handling releasing of memory
         private readonly Action<T> m_objectDeallocator;
-        // Object storage
+        //! Object storage
         private readonly List<T> m_objects;
-        // Index to the first available object in object pool
+        //! Index to the first available object in object pool
         private int m_objectIndex;
-        // Initial size of object pool. We never deallocate memory under tris threshold
-        private int m_initialSize;
+        //! Initial size of object pool. We never deallocate memory under this threshold
+        private readonly int m_initialSize;
+        //! If true object pool will try to release some of the unused memory if the difference in currently used size and capacity of pool is too big
+        private readonly bool m_autoReleaseMemory;
 
-        public ObjectPool(Func<T, T> objectAllocator, int initialSize)
+        public ObjectPool(Func<T, T> objectAllocator, int initialSize, bool autoReleaseMememory)
         {
             m_objectAllocator = new ObjectPoolAllocator<T>(objectAllocator);
             m_objectDeallocator = null;
             m_initialSize = initialSize;
+            m_autoReleaseMemory = autoReleaseMememory;
             m_objectIndex = 0;
 
             m_objects = new List<T>(initialSize);
@@ -29,11 +32,12 @@ namespace Assets.Engine.Scripts.Common.Collections
                 m_objects.Add(m_objectAllocator.Action(m_objectAllocator.Arg));
         }
 
-        public ObjectPool(ObjectPoolAllocator<T> objectAllocator, int initialSize)
+        public ObjectPool(ObjectPoolAllocator<T> objectAllocator, int initialSize, bool autoReleaseMememory)
         {
             m_objectAllocator = objectAllocator;
             m_objectDeallocator = null;
             m_initialSize = initialSize;
+            m_autoReleaseMemory = autoReleaseMememory;
             m_objectIndex = 0;
 
             m_objects = new List<T>(initialSize);
@@ -46,6 +50,7 @@ namespace Assets.Engine.Scripts.Common.Collections
             m_objectAllocator = new ObjectPoolAllocator<T>(objectAllocator);
             m_objectDeallocator = objectDeallocator;
             m_initialSize = initialSize;
+            m_autoReleaseMemory = true;
             m_objectIndex = 0;
 
             m_objects = new List<T>(initialSize);
@@ -58,6 +63,7 @@ namespace Assets.Engine.Scripts.Common.Collections
             m_objectAllocator = objectAllocator;
             m_objectDeallocator = objectDeallocator;
             m_initialSize = initialSize;
+            m_autoReleaseMemory = true;
             m_objectIndex = 0;
 
             m_objects = new List<T>(initialSize);
@@ -89,20 +95,23 @@ namespace Assets.Engine.Scripts.Common.Collections
             // If we're using less then 1/4th of memory capacity, let's free half of the allocated memory.
             // We're doing it this way so that there's a certain threshold before allocating new memory.
             // We only deallocate if there's at least m_initialSize items allocated.
-            int thresholdCount = m_objects.Count>>2;
-            if (thresholdCount > m_initialSize && m_objectIndex <= thresholdCount)
+            if (m_autoReleaseMemory)
             {
-                int halfCount = m_objects.Count>>1;
-
-                // Use custom deallocation if deallocator is set
-                if (m_objectDeallocator!=null)
+                int thresholdCount = m_objects.Count>>2;
+                if (thresholdCount>m_initialSize && m_objectIndex<=thresholdCount)
                 {
-                    for (int i = halfCount; i<m_objects.Count; i++)
-                        m_objectDeallocator(m_objects[i]);
-                }
+                    int halfCount = m_objects.Count>>1;
 
-                // Remove one half of unused items
-                m_objects.RemoveRange(halfCount, halfCount);
+                    // Use custom deallocation if deallocator is set
+                    if (m_objectDeallocator!=null)
+                    {
+                        for (int i = halfCount; i<m_objects.Count; i++)
+                            m_objectDeallocator(m_objects[i]);
+                    }
+
+                    // Remove one half of unused items
+                    m_objects.RemoveRange(halfCount, halfCount);
+                }
             }
 
             m_objects[--m_objectIndex] = item;
